# swagger.yaml
openapi: 3.0.3
info:
  title: BackendLab API
  description: |
    API educativa con tres laboratorios independientes sobre NestJS + TypeORM.

    **Lab 1 – Fundamentos**: Params, Query, Body, Headers, HttpCode y CRUD in-memory de productos.
    **Lab 2 – Patrones**: Pipes, Guards, Interceptors, Exception Filters, Custom Decorators, Middleware.
    **Lab 3 – Integraciones**: JWT Auth (PostgreSQL/TypeORM), Posts CRUD, File Upload, EventEmitter.
    **Northwind**: Endpoints de lectura sobre las 14 tablas de la base Northwind en Supabase PostgreSQL.

    ### Quick Start – Lab 3
    1. `POST /lab3/auth/login` con `{ "email": "admin@lab3.com", "password": "admin123" }`
    2. Copiá el `access_token` y usalo en el header `Authorization: Bearer <token>`
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido desde POST /lab3/auth/login
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: "Recurso no encontrado"
        error:
          type: string
          example: "Not Found"
...
    Lab1Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Laptop Pro 15"
        price:
          type: number
          format: double
          example: 999.99
        category:
          type: string
          example: "Electronics"
        stock:
          type: integer
          example: 100
...
    Lab1UpdateProductDto:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
          example: "Laptop Pro 15 (actualizado)"
        price:
          type: number
          format: double
          minimum: 0.01
          example: 799.99
        category:
          type: string
          example: "Electronics"
        stock:
          type: integer
          minimum: 0
          example: 50
...
    NorthwindCustomer:
      type: object
      properties:
        customerId:
          type: string
          description: Clave natural de 5 caracteres (ej. "ALFKI")
          example: "ALFKI"
        companyName:
          type: string
          example: "Alfreds Futterkiste"
        contactName:
          type: string
          nullable: true
          example: "Maria Anders"
        contactTitle:
          type: string
          nullable: true
          example: "Sales Representative"
        address:
          type: string
          nullable: true
          example: "Obere Str. 57"
        city:
          type: string
          nullable: true
          example: "Berlin"
        region:
          type: string
          nullable: true
          example: null
        postalCode:
          type: string
          nullable: true
          example: "12209"
        country:
          type: string
          nullable: true
          example: "Germany"
        phone:
          type: string
          nullable: true
          example: "030-0074321"
        fax:
          type: string
          nullable: true
          example: "030-0076545"
    NorthwindProduct:
      type: object
      properties:
        productId:
          type: integer
          example: 1
        productName:
          type: string
          example: "Chai"
        supplierId:
          type: integer
          nullable: true
          example: 1
        categoryId:
          type: integer
          nullable: true
          example: 1
        quantityPerUnit:
          type: string
          nullable: true
          example: "10 boxes x 20 bags"
        unitPrice:
          type: number
          format: double
          nullable: true
          example: 18.00
        unitsInStock:
          type: integer
          nullable: true
          example: 39
        unitsOnOrder:
          type: integer
          nullable: true
          example: 0
        reorderLevel:
          type: integer
          nullable: true
          example: 10
        discontinued:
          type: integer
          description: "0 = activo, 1 = descontinuado"
          example: 0
    NorthwindOrder:
      type: object
      properties:
        orderId:
          type: integer
          example: 10248
        customerId:
          type: string
          nullable: true
          example: "VINET"
        employeeId:
          type: integer
          nullable: true
          example: 5
        orderDate:
          type: string
          format: date
          nullable: true
          example: "1996-07-04"
        requiredDate:
          type: string
          format: date
          nullable: true
          example: "1996-08-01"
        shippedDate:
          type: string
          format: date
          nullable: true
          example: "1996-07-16"
        shipVia:
          type: integer
          nullable: true
          description: FK hacia shippers.shipperId
          example: 3
        freight:
          type: number
          format: double
          nullable: true
          example: 32.38
        shipName:
          type: string
          nullable: true
          example: "Vins et alcools Chevalier"
        shipAddress:
          type: string
          nullable: true
          example: "59 rue de l'Abbaye"
        shipCity:
          type: string
          nullable: true
          example: "Reims"
        shipRegion:
          type: string
          nullable: true
          example: null
        shipPostalCode:
          type: string
          nullable: true
          example: "51100"
        shipCountry:
          type: string
          nullable: true
          example: "France"
...
paths:
...
  /lab1/products/{id}:
    get:
      tags: [Lab 1 – Products]
      summary: Obtener producto por ID
      operationId: lab1FindOneProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Producto encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Lab1Product'
        "404":
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags: [Lab 1 – Products]
      summary: Actualizar producto parcialmente (PATCH)
      description: Solo enviá los campos que querés modificar. PATCH es actualización parcial, no reemplaza el recurso completo (que sería PUT).
      operationId: lab1UpdateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lab1UpdateProductDto'
            example:
              price: 799.99
              stock: 50
      responses:
        "200":
          description: Producto actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Lab1Product'
                  message:
                    type: string
                    example: "Product updated successfully"
        "404":
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Lab 1 – Products]
      summary: Eliminar producto
      operationId: lab1DeleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Producto eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Lab1Product'
                  message:
                    type: string
                    example: "Product deleted successfully"
        "404":
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ═══════════════════════════════════════════════════════════════════════════
  # LAB 2 – PIPES
  # ═══════════════════════════════════════════════════════════════════════════
...
  /lab2/pipes/custom/{value}:
    get:
      tags: [Lab 2 – Pipes]
      summary: CustomUppercasePipe — transforma a MAYÚSCULAS
      description: Pipe personalizado que toma el string del param y lo convierte a uppercase antes de que llegue al handler.
      operationId: lab2CustomPipe
      parameters:
        - name: value
          in: path
          required: true
          schema:
            type: string
          example: "hola-mundo"
      responses:
        "200":
          description: String transformado a UPPERCASE
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      value:
                        type: string
                        example: "HOLA-MUNDO"
                      note:
                        type: string
                        example: "Transformed to UPPERCASE by CustomUppercasePipe"
...
  /customers/search:
    get:
      tags: [Northwind – Customers]
      summary: Buscar clientes por nombre de empresa
      operationId: northwindSearchCustomers
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          example: "Alfreds"
      responses:
        "200":
          description: Clientes que coinciden con el nombre
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NorthwindCustomer'
  /customers/country/{country}:
    get:
      tags: [Northwind – Customers]
      summary: Filtrar clientes por país
      operationId: northwindGetCustomersByCountry
      parameters:
        - name: country
          in: path
          required: true
          schema:
            type: string
          example: "Germany"
      responses:
        "200":
          description: Clientes del país indicado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NorthwindCustomer'
  /customers/{id}:
    get:
      tags: [Northwind – Customers]
      summary: Obtener cliente por ID (customer_id varchar)
      operationId: northwindGetCustomerById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "ALFKI"
      responses:
        "200":
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NorthwindCustomer'
        "404":
          description: Cliente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /products/search:
    get:
      tags: [Northwind – Products]
      summary: Buscar productos por nombre
      operationId: northwindSearchProducts
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          example: "Chai"
      responses:
        "200":
          description: Productos que coinciden con el nombre
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NorthwindProduct'
  /products/category/{categoryId}:
    get:
      tags: [Northwind – Products]
      summary: Filtrar productos por categoría
      operationId: northwindGetProductsByCategory
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Productos de la categoría
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NorthwindProduct'
  /products/supplier/{supplierId}:
    get:
      tags: [Northwind – Products]
      summary: Filtrar productos por proveedor
      operationId: northwindGetProductsBySupplier
      parameters:
        - name: supplierId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Productos del proveedor
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NorthwindProduct'
  /products/{id}:
    get:
      tags: [Northwind – Products]
      summary: Obtener producto por ID
      operationId: northwindGetProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NorthwindProduct'
        "404":
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /orders/customer/{customerId}:
    get:
      tags: [Northwind – Orders]
      summary: Listar órdenes de un cliente
      operationId: northwindGetOrdersByCustomer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
          example: "ALFKI"
      responses:
        "200":
          description: Órdenes del cliente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NorthwindOrder'
  /orders/employee/{employeeId}:
    get:
      tags: [Northwind – Orders]
      summary: Listar órdenes atendidas por un empleado
      operationId: northwindGetOrdersByEmployee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Órdenes del empleado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NorthwindOrder'
  /orders/{id}:
    get:
      tags: [Northwind – Orders]
      summary: Obtener orden por ID
      operationId: northwindGetOrderById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 10248
      responses:
        "200":
          description: Orden encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NorthwindOrder'
        "404":
          description: Orden no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...